+++ frontend/src/views/Collection/index.tsx
@@ -1,4 +1,4 @@
-import { Button, Classes, H3, H6, Intent } from "@blueprintjs/core";
+import { Button, Classes, H3, Intent } from "@blueprintjs/core";
 import { IconNames } from "@blueprintjs/icons";
 import { RouteComponentProps } from "@reach/router";
 import { memoize } from "lodash-es";
@@ -14,14 +14,13 @@ import { hasAssets } from "src/common/modules/datasets/selectors";
 import {
   useCollection,
   useCollectionUploadLinks,
-  useDeleteCollection,
   USE_COLLECTION,
 } from "src/common/queries/collections";
 import { getUrlHost } from "src/common/utils/getUrlHost";
+import DeleteCollection from "src/components/Collections/components/Collection/common/components/DeleteCollection";
 import DownloadDataset from "src/components/Collections/components/Dataset/components/DownloadDataset";
 import DatasetsGrid from "src/components/Collections/components/Grid/components/DatasetsGrid";
 import DeleteDataset from "src/components/Collections/components/Grid/components/Row/DatasetRow/components/DeleteDataset";
-import { StyledAlert } from "src/components/Collections/components/Grid/components/Row/DatasetRow/components/DeleteDataset/style";
 import DropboxChooser, { UploadingFile } from "src/components/DropboxChooser";
 import { ViewGrid } from "../globalStyle";
 import { StyledLink } from "./common/style";
@@ -104,18 +103,6 @@ const Collection: FC<Props> = ({ id = "" }) => {
     );
   };
 
-  const [deleteMutation] = useDeleteCollection(id);
-
-  const handleDelete = () => {
-    deleteMutation(id);
-  };
-
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleAlert = () => {
-    setIsOpen(!isOpen);
-  };
-
   if (!collection || isError) {
     return null;
   }
@@ -144,32 +131,7 @@ const Collection: FC<Props> = ({ id = "" }) => {
       </CollectionInfo>
 
       <CollectionButtons>
-        <Button
-          intent={Intent.DANGER}
-          minimal
-          text="Delete"
-          icon={IconNames.TRASH}
-          onClick={toggleAlert}
-        />
-
-        <StyledAlert
-          cancelButtonText={"Cancel"}
-          confirmButtonText={"Delete Collection"}
-          intent={Intent.DANGER}
-          isOpen={isOpen}
-          onCancel={toggleAlert}
-          onConfirm={() => {
-            handleDelete();
-            toggleAlert();
-          }}
-        >
-          <H6>Are you sure you want to delete this collection?</H6>
-          <p>
-            Deleting this collection will also delete any uploaded datasets. If
-            you’ve shared this collection or its datasets with anyone, they will
-            also lose access. You cannot undo this action.
-          </p>
-        </StyledAlert>
+        <DeleteCollection id={id} />
         <Button
           intent={Intent.PRIMARY}
           minimal
